+++ src/api/endpoints/tokens/get-tokens-details/v2.ts
@@ -370,14 +370,11 @@ export const getTokensDetailsV2Options: RouteOptions = {
         continuation = buildContinuation(continuation);
       }
 
-      const result = rawResult.map(async (r) => {
-        const sources = new Sources();
+      const sources = await Sources.getInstance();
+
+      const result = rawResult.map((r) => {
         const source = r.floor_sell_source_id
-          ? await sources.get(
-              fromBuffer(r.floor_sell_source_id),
-              fromBuffer(r.contract),
-              r.token_id
-            )
+          ? sources.get(fromBuffer(r.floor_sell_source_id), fromBuffer(r.contract), r.token_id)
           : null;
 
         return {
@@ -428,7 +425,7 @@ export const getTokensDetailsV2Options: RouteOptions = {
       });
 
       return {
-        tokens: await Promise.all(result),
+        tokens: result,
         continuation,
       };
     } catch (error) {