+++ src/api/endpoints/tokens/get-tokens-details/v3.ts
@@ -402,15 +402,11 @@ export const getTokensDetailsV3Options: RouteOptions = {
 
         continuation = buildContinuation(continuation);
       }
+      const sources = await Sources.getInstance();
 
-      const result = rawResult.map(async (r) => {
-        const sources = new Sources();
+      const result = rawResult.map((r) => {
         const source = r.floor_sell_source_id
-          ? await sources.get(
-              fromBuffer(r.floor_sell_source_id),
-              fromBuffer(r.contract),
-              r.token_id
-            )
+          ? sources.get(fromBuffer(r.floor_sell_source_id), fromBuffer(r.contract), r.token_id)
           : null;
 
         return {
@@ -463,7 +459,7 @@ export const getTokensDetailsV3Options: RouteOptions = {
       });
 
       return {
-        tokens: await Promise.all(result),
+        tokens: result,
         continuation,
       };
     } catch (error) {