+++ src/models/sources/index.ts
@@ -1,10 +1,39 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import _ from "lodash";
 
 import { config } from "@/config/index";
 import { idb } from "@/common/db";
 import { SourcesEntity, SourcesEntityParams } from "@/models/sources/sources-entity";
 
+import { default as sources } from "./sources.json";
+
 export class Sources {
+  private static instance: Sources;
+
+  public sources: object;
+
+  // eslint-disable-next-line @typescript-eslint/no-empty-function
+  private constructor() {
+    this.sources = {};
+  }
+
+  private async loadData() {
+    const sources: SourcesEntityParams[] | null = await idb.manyOrNone(`SELECT * FROM sources`);
+
+    for (const source of sources) {
+      (this.sources as any)[source.source_id] = new SourcesEntity(source);
+    }
+  }
+
+  public static async getInstance() {
+    if (!this.instance) {
+      this.instance = new Sources();
+      await this.instance.loadData();
+    }
+
+    return this.instance;
+  }
+
   public static getDefaultSource(sourceId: string): SourcesEntity {
     return new SourcesEntity({
       source_id: sourceId,
@@ -18,34 +47,33 @@ export class Sources {
     });
   }
 
-  public async getAll() {
-    const sources: SourcesEntityParams[] | null = await idb.manyOrNone(
-      `SELECT *
-             FROM sources`
-    );
-
-    if (sources) {
-      return _.map(sources, (source) => new SourcesEntity(source));
-    }
-
-    return null;
+  public static async syncSources() {
+    _.forEach(sources, (metadata, sourceId) => {
+      Sources.add(sourceId, metadata);
+    });
   }
 
-  public async get(sourceId: string, contract?: string, tokenId?: string) {
-    let sourceEntity;
-    const source: SourcesEntityParams | null = await idb.oneOrNone(
-      `SELECT *
-              FROM sources
-              WHERE source_id = $/sourceId/`,
-      {
-        sourceId,
-      }
-    );
+  public static async add(sourceId: string, metadata: object) {
+    const query = `INSERT INTO sources (source_id, metadata)
+                   VALUES ( $/sourceId/, $/metadata:json/)
+                   ON CONFLICT (source_id) DO UPDATE
+                   SET metadata = $/metadata:json/`;
 
-    if (!source) {
-      sourceEntity = Sources.getDefaultSource(sourceId);
+    const values = {
+      sourceId,
+      metadata: metadata,
+    };
+
+    await idb.none(query, values);
+  }
+
+  public get(sourceId: string, contract?: string, tokenId?: string) {
+    let sourceEntity;
+
+    if (sourceId in this.sources) {
+      sourceEntity = (this.sources as any)[sourceId];
     } else {
-      sourceEntity = new SourcesEntity(source);
+      sourceEntity = Sources.getDefaultSource(sourceId);
     }
 
     if (config.chainId == 1) {