+++ src/models/tokens/index.ts
@@ -5,8 +5,15 @@ import { toBuffer } from "@/common/utils";
 import { TokensEntity, TokensEntityParams, TokensEntityUpdateParams } from "./tokens-entity";
 import _ from "lodash";
 
+export type TokenAttributes = {
+  attributeId: number;
+  key: string;
+  value: string;
+  attributeKeyId: number;
+};
+
 export class Tokens {
-  public static async getByContractAndTokenId(contract: string, tokenId: number) {
+  public static async getByContractAndTokenId(contract: string, tokenId: string) {
     const token: TokensEntityParams | null = await idb.oneOrNone(
       `SELECT *
               FROM tokens
@@ -25,7 +32,7 @@ export class Tokens {
     return null;
   }
 
-  public static async update(contract: string, tokenId: number, fields: TokensEntityUpdateParams) {
+  public static async update(contract: string, tokenId: string, fields: TokensEntityUpdateParams) {
     let updateString = "";
     const replacementValues = {
       contract: toBuffer(contract),
@@ -45,5 +52,18 @@ export class Tokens {
                    AND token_id = $/tokenId/`;
 
     return await idb.none(query, replacementValues);
+  }
+
+  public static async getTokenAttributes(contract: string, tokenId: string) {
+    const query = `SELECT attribute_id AS attributeId, key, token_attributes.value, attribute_key_id AS attributeKeyId
+                   FROM token_attributes
+                   JOIN attributes ON token_attributes.attribute_id = attributes.id
+                   WHERE contract = $/contract/
+                   AND token_id = $/tokenId/`;
+
+    return (await idb.manyOrNone(query, {
+      contract: toBuffer(contract),
+      tokenId,
+    })) as TokenAttributes[];
   }
 }